[refactor] 페이지 분기 로직 제거(CUL-243)
