+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/performance/entity/Sort.java
@@ -1,4 +1,4 @@
-package com.culturespot.culturespotserviceapi.common.dto;
+package com.culturespot.culturespotdomain.core.performance.entity;
 
 import lombok.Getter;
