+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/performance/repository/PerformanceRepository.java
@@ -1,13 +1,14 @@
 package com.culturespot.culturespotdomain.core.performance.repository;
 
+import com.culturespot.culturespotdomain.core.performance.entity.Category;
 import com.culturespot.culturespotdomain.core.performance.entity.Event;
 import com.culturespot.culturespotdomain.core.performance.entity.Performance;
-
 import java.util.Collection;
 import java.util.List;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -17,10 +18,16 @@ public interface PerformanceRepository extends JpaRepository<Performance, Long> 
   List<Performance> findBySeqIn(Collection<String> seqValues);
 
   @Query("""
-          SELECT p FROM Performance p
-          WHERE p.type = :event
-          ORDER BY p.createdAt DESC
-          """)
-  Page<Performance> findMainLatestPerformances(@Param("event") Event event, Pageable pageable);
-
+    SELECT p FROM Performance p
+    WHERE (:event IS NULL OR p.type = :event)
+      AND (:category IS NULL OR p.category = :category)
+      AND (:keyword IS NULL OR LOWER(p.title) LIKE LOWER(CONCAT('%', :keyword, '%')))
+      AND (:lastId IS NULL OR p.id < :lastId)
+    ORDER BY p.createdAt DESC
+""")
+  Page<Performance> findLatestPerformances(@Param("event") Event event,
+                                     @Param("category") Category category,
+                                     @Param("keyword") String keyword,
+                                     @Param("lastId") Long lastId,
+                                     Pageable pageable);
 }
